Moved temperature dependent sigma to parameters so that copies of HardSphere and Association are no longer needed.
